fix: delete user_info model and add column name twitter_id instagram_id to user
