+++ app/controllers/users_controller.rb
@@ -1,13 +1,12 @@
 class UsersController < ApplicationController
   before_action :set_user, only: %i[ show edit update ]
   before_action :sign_in_required, only: [:edit, :update]
-  before_action :correct_user,   only: [:edit, :update]
+  before_action :correct_user, only: [:edit, :update]
 
   def show
   end
 
   def edit
-    @user = @user || current_user.create_user_info
   end
 
   def update
@@ -22,11 +21,11 @@ class UsersController < ApplicationController
 
   private
     def user_params
-      params.require(:user_info).permit(:name, :twitter_id, :instagram_id)
+      params.require(:user).permit(:name, :twitter_id, :instagram_id)
     end
 
     def set_user
-      @user = User.find(params[:id]).user_info
+      @user = User.find(params[:id])
     end
 
     def correct_user
