+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2022_04_12_153226) do
+ActiveRecord::Schema.define(version: 2022_04_13_153237) do
 
   create_table "active_storage_attachments", charset: "utf8mb4", collation: "utf8mb4_0900_ai_ci", force: :cascade do |t|
     t.string "name", null: false
@@ -81,16 +81,6 @@ ActiveRecord::Schema.define(version: 2022_04_12_153226) do
     t.datetime "updated_at", precision: 6, null: false
   end
 
-  create_table "user_infos", charset: "utf8mb4", collation: "utf8mb4_0900_ai_ci", force: :cascade do |t|
-    t.bigint "user_id", null: false
-    t.string "name", default: ""
-    t.string "twitter_id", default: ""
-    t.string "instagram_id", default: ""
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
-    t.index ["user_id"], name: "index_user_infos_on_user_id"
-  end
-
   create_table "users", charset: "utf8mb4", collation: "utf8mb4_0900_ai_ci", force: :cascade do |t|
     t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
@@ -99,6 +89,9 @@ ActiveRecord::Schema.define(version: 2022_04_12_153226) do
     t.datetime "remember_created_at"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.string "name", default: ""
+    t.string "twitter_id", default: ""
+    t.string "instagram_id", default: ""
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
   end
@@ -109,5 +102,4 @@ ActiveRecord::Schema.define(version: 2022_04_12_153226) do
   add_foreign_key "circle_images", "circles"
   add_foreign_key "circle_images", "images"
   add_foreign_key "circle_times", "circles"
-  add_foreign_key "user_infos", "users"
 end
