+++ spec/factories/users.rb
@@ -2,8 +2,8 @@ FactoryBot.define do
   factory :user do
     email { Faker::Internet.email }
     password { Faker::Internet.password }
-    after(:build) do |user|
-      user.user_info || build(:user_info, user: user)
-    end
+    name { Faker::Internet.username }
+    twitter_id { Faker::Lorem.characters(number: 10) }
+    instagram_id { Faker::Lorem.characters(number: 10) }
   end
 end
